Conflicts:
	Labeler.m
	trackers/dt/BgWorkerObjLocalFilesys.m
	trackers/dt/DeepTracker.m

+++ deepnet/open_pose.py
@@ -646,7 +646,7 @@ def training(conf,name='deepnet'):
                 json.dump(json_data, json_file)
 
             if step % conf.save_step == 0:
-                model.save(str(os.path.join(conf.cachedir, name + '-{}'.format(step))))
+                model.save(str(os.path.join(conf.cachedir, name + '-{}'.format(int(step)))))
 
 
     # configure callbacks
@@ -677,7 +677,7 @@ def training(conf,name='deepnet'):
                         )
 
     # force saving in case the max iter doesn't match the save step.
-    model.save(str(os.path.join(conf.cachedir, name + '-{}'.format(max_iter*iterations_per_epoch))))
+    model.save(str(os.path.join(conf.cachedir, name + '-{}'.format(int(max_iter*iterations_per_epoch)))))
     obs.on_epoch_end(max_iter-1)
